+++ src/DataSource/AccountDataSourceContext.ts
@@ -1,14 +1,14 @@
 import { DataSourceManager } from './DataSourceManager';
 import { AR, OKA } from '@hexancore/common';
 
-import { DataSourceContext } from './DataSourceContext';
+import { AbstractDataSourceContext } from './AbstractDataSourceContext';
 import { DataSourceContextConfig } from './DataSourceContextConfig';
 import { Injectable } from '@nestjs/common';
 import { AccountContext } from '@hexancore/core';
 import { TYPEORM_ACCOUNT_PERSISTER_TYPE } from '@/Repository';
 
 @Injectable()
-export class AccountDataSourceContext extends DataSourceContext {
+export class AccountDataSourceContext extends AbstractDataSourceContext {
   public constructor(
     manager: DataSourceManager,
     protected ac: AccountContext,