+++ test/helper/src/Test/Infrastructure/Persistence/TypeOrm/Account/Author/TypeOrmBookRepository.ts
@@ -11,7 +11,7 @@ export const BookSchema = TypeOrmAccountEntitySchema(Book, {
   columns: {
     id: UIntValueColumn.asPrimaryKey(BookId),
     authorId: UIntValueColumn.as(AuthorId, { type: 'smallint' }),
-    name: { type: 'varchar', length: 255 },
+    title: { type: 'varchar', length: 255 },
     createdAt: DateTimeColumn.asSelf(),
   },
 });