+++ test/helper/src/Test/Infrastructure/PrivateTestInfraModule.ts
@@ -3,5 +3,6 @@ import { Module } from '@nestjs/common';
 
 @Module({
   imports: [HcInfraDomainModule.forFeature({ moduleInfraDir: __dirname })],
+  exports: [HcInfraDomainModule],
 })
 export class PrivateTestInfraModule {}