+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.15.0] - 2024-06-24
+
+### Changed
+
+- Updated @hexancore/common to 0.14.*
+- Updated @hexancore/core to  ^0.15.*
+
+### Removed
+
+- removed support for mysql driver
+
+
 ## [0.14.0] - 2024-02-11
 
 ### Changed
@@ -59,7 +71,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - First release(start from 0.12 for consistency)
 - Implemented generic from domain infrastructre from Hexancore
 
-[unreleased] https://github.com/hexancore/typeorm/compare/0.14.0...HEAD   
+[unreleased] https://github.com/hexancore/typeorm/compare/0.15.0...HEAD   
+[0.15.0] https://github.com/hexancore/typeorm/compare/0.14.0...0.15.0   
 [0.14.0] https://github.com/hexancore/typeorm/compare/0.13.3...0.14.0   
 [0.13.3] https://github.com/hexancore/typeorm/compare/0.13.2...0.13.3   
 [0.13.2] https://github.com/hexancore/typeorm/compare/0.13.1...0.13.2